platforms/nuttx/src/bootloader/common/memtest.c: Change datum to 32 bit data type
The data bus is 32-bits wide, but on this platform unsigned long is 64-bits.
Change to "unsigned"

+++ platforms/nuttx/src/bootloader/common/memtest.h
@@ -28,7 +28,7 @@
 /*
  * Set the data bus width.
  */
-typedef unsigned long datum;
+typedef unsigned datum;
 
 /*
  * Function prototypes.
